Fix `finalize` and `Context.finalize` functions
These finalize functions are expected to return false only if no
cleaning occurred between evaluation of the `doGC` function argument and
the function returning, to indicate that further GC would not find any
other unreachable finalizable values, but this was not guaranteed [1]:
after one of these finalize functions evaluated its `doGC` function
argument, another thread (Poly/ML) or handler (MLton) could clean some
pending lists so the finalize function performed no further cleaning and
returned false, even though further GC could make more finalizable
values reachable, enabling more finalizers to be run.  This issue is
fixed by introducing a list of notifiers with each pending list.  A
notifier is a `bool ref`.  When a pending list is cleaned, its notifiers
are all set to true.  A finalize function adds a notifier before
evaluating `doGC` and removes the notifier after cleaning, returning the
notifier value.  Therefore a finalize function is guaranteed to detect
any cleaning between evaluation of the `doGC` function argument and the
returning.  Note that cleaning outside this window may be detected
causing the finalize function to return true when it could return false
but this does not matter and is unlikely.

Another issue with these functions is that they caused the next round of
cleaning by the GC handler to be skipped, to avoid double-cleaning, but
the `doGC` function argument may not invoke GC: it could simply be the
identity function.  Therefore cleaning after a subsequent GC would not
occur when it should.  The issue is avoided by using a different method
to reduce the likelihood of double-cleaning: cleaning in the GC handler
cleans a pending list only if it has no notifiers.

 1. As pointed out by @MatthewFluet
    https://github.com/MLton/mlton/issues/508#issuecomment-1685492326