Change `GLib.Source.t` constructor functions to take callback function
In C, `g_source_set_callback` is used to set the callback function of a
source after it has been created.  The type of the callback function
varies so this API cannot be carried over to SML whilst maintaining type
safety.  One option is to generalize the type of a source by introducing
a type parameter for the type of its callback function.  However, such
a representation is not supported by code generation and would require
significant manual intervention in the generated code, introducing a
significant maintenance overhead.  Instead, the SML functions that
construct a new kind of source are bindings to extended versions of the
C functions that also take the callback function for the source.  The
signature G_LIB now specifies:

    val childWatchSourceNew : Pid.t * ChildWatchFunc.t -> SourceRecord.t
    val idleSourceNew : SourceFunc.t -> SourceRecord.t
    val timeoutSourceNew : LargeInt.int * SourceFunc.t -> SourceRecord.t
    val timeoutSourceNewSeconds
                         : LargeInt.int * SourceFunc.t -> SourceRecord.t

These functions were effectively useless because a callback could not be
set after a source was created, so this change should not affect
existing code.