Change SML signal handler functions to have instance object argument
The first argument of the handler function of a signal is now (a fresh
reference to) the object that emitted the signal.  Although this is not
compatible with existing code, the required changes to existing code are
likely to identify cases where otherwise unused objects are prevented
from being destroyed due to a cyclic dependency between an object and
its handler function.

Using the handler argument to refer to the object that emitted a signal
means that the object is referenced by the handler function only when
the signal is emitted, and the reference can be dropped subsequently.
If the first argument is not used, as was the case before this change,
and a handler function holds a reference to the object directly, the
object cannot be destroyed until the handler function is explicitly
disconnected.