Use thread-local finalization for objects
A finalizer added using the structure Finalizable is run in either its
own OS thread (Poly/ML) or asynchronously in an interrupt (MLton).  This
is not suitable for an object's finalizer because thread-safety of an
object is class-dependent.  Notably, GTK classes provide no guarantee of
thread-safety.  In the absence of a general way to determine whether a
class is thread-safe, it is assumed that none are.  Therefore, object
finalizers are added using the structure ThreadFinalizable to ensure
that they are run from the thread in which the object was created.

A new substructure Finalize is added to the structure Giraffe that wraps
ThreadFinalizable to give an interface suitable for applications. This
provides functions for performing finalization both synchronously and
asynchronously.  Asynchronous finalization is achieved by setting up a
marshaler that schedules finalization in a specified main context on GC.

ThreadFinalizable is updated to set marshalers for all threads to
`marshalNow` at exit so that thread-local finalizers are run before an
application exits.  This assumes that the thread running the at-exit
functions is the only thread left running.

ContextFinalizable is updated to force finalization at exit, if enabled,
of contexts without a marshaler, by using `marshalNow`.  This behaviour
is preferable because forced finalization is used for debugging purposes
only and is intended to run as many blocked finalizers as possible,
regardless of whether it is valid to do so.